Add CSS animations and color to the site
Added simple CSS animations and color changes to the site to make it look better. The animations were a big learning curve as CSS doesn't operate like Bash or Powershell like I'm used to. It's strange to see the animation defined at the bottom of the file but still be able to use it earlier.